This project was upgraded with external_updater.
Usage: tools/external_updater/updater.sh update bcc
For more info, check https://cs.android.com/android/platform/superproject/+/main:tools/external_updater/README.md

Conflicts resolved:
    src/cc/libbpf.c

Resolved build issues:
    Android.bp
    libbpf-tools/Android.bp
    src/cc/libbpf.c - wrap find_btf_id in "#ifndef MINIMAL_LIBBPF"
    src/cc/bcc_zip.c - pointer casts to resolve -Wgnu-pointer-arith and
                         -Wcompare-distinct-pointer-types build errors

Test: TreeHugger

+++ tools/dbslower.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/env python
 #
 # dbslower      Trace MySQL and PostgreSQL queries slower than a threshold.
 #
@@ -100,7 +100,7 @@ struct temp_t {
 };
 
 struct data_t {
-    u64 pid;
+    u32 pid;
     u64 timestamp;
     u64 duration;
     char query[256];
